pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "simple-agent"
version = "0.1.0"
description = "A simple conversational AI agent demonstrating the Observe-Think-Act pattern"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "anthropic>=0.39.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=3.0.0",
    "black>=24.0.0",
    "flake8>=7.0.0",
    "pytest>=8.0.0",
]

[project.scripts]
simple-agent = "simple_agent.simple:main"
advanced-agent = "simple_agent.advanced:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ["py39", "py310", "py311", "py312", "py313"]

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]

[tool.pytest.ini_options]
testpaths = ["tests"]
